FOPTS-2404 Save 2 months of csv files. Validate number of files to upload for previous month (#27)
* FOPTS-2404 Save 2 months of csv files. Validate number of files to upload for previous month

* FOPTS-2404 Removed new env var SEND_ONLY_FULL_PREVIOUS_MONTH to don't overcomplicate configuration

+++ main.go
@@ -124,11 +124,12 @@ type (
 
 	App struct {
 		Config
-		aggregation     string
-		filesToUpload   map[string]map[string]struct{}
-		client          *http.Client
-		lastInvoiceDate time.Time
-		invoiceMonths   []string
+		aggregation                        string
+		filesToUpload                      map[string]map[string]struct{}
+		client                             *http.Client
+		lastInvoiceDate                    time.Time
+		invoiceMonths                      []string
+		mandatoryFileSavingPeriodStartDate time.Time
 	}
 )
 
@@ -198,6 +199,11 @@ func (a *App) updateFromKubecost() {
 		monthOfData := d.Format("2006-01")
 		var csvFile = fmt.Sprintf(path.Join(a.FilePath, "kubecost-%v.csv"), d.Format("2006-01-02"))
 
+		if j.Code != http.StatusOK {
+			log.Println("Kubecost API response code different than 200, skipping")
+			continue
+		}
+
 		// If the data obtained is empty, skip the iteration, because it might overwrite a previously obtained file for the same range time
 		_, previousFileCreated := a.filesToUpload[monthOfData][csvFile]
 		if len(data) == 0 && previousFileCreated {
@@ -243,6 +249,18 @@ func (a *App) uploadToFlexera() {
 	billUploadURL := fmt.Sprintf("https://%s/optima/orgs/%s/billUploads", shardDict[a.Shard], a.OrgID)
 
 	for month, files := range a.filesToUpload {
+
+		if len(files) == 0 {
+			log.Println("No files to upload for month", month)
+			continue
+		}
+
+		// if we try to upload files for previous month, we need to check if we have files for all days in the month
+		if !a.isCurrentMonth(month) && a.DaysInMonth(month) != len(files) {
+			log.Println("Skipping month", month, "because not all days have a file to upload")
+			continue
+		}
+
 		authHeaders := map[string]string{"Authorization": "Bearer " + accessToken}
 		billUpload := map[string]string{"billConnectId": a.BillConnectID, "billingPeriod": month}
 
@@ -374,7 +392,7 @@ func (a *App) updateFileList() {
 			if t, err := time.Parse("kubecost-2006-01-02.csv", file.Name()); err == nil {
 				if a.dateInInvoiceRange(t) {
 					a.filesToUpload[t.Format("2006-01")][path.Join(a.FilePath, file.Name())] = struct{}{}
-				} else if a.FileRotation {
+				} else if a.FileRotation && !a.dateInMandatoryFileSavingPeriod(t) {
 					if err = os.Remove(path.Join(a.FilePath, file.Name())); err != nil {
 						log.Printf("error removing file %s: %v", file.Name(), err)
 					}
@@ -420,6 +438,23 @@ func (a *App) dateInInvoiceRange(date time.Time) bool {
 	return false
 }
 
+func (a *App) dateInMandatoryFileSavingPeriod(date time.Time) bool {
+	return !date.Before(a.mandatoryFileSavingPeriodStartDate)
+}
+
+func (a *App) isCurrentMonth(month string) bool {
+	return time.Now().Local().Format("2006-01") == month
+}
+
+func (a *App) DaysInMonth(month string) int {
+	date, err := time.Parse("2006-01", month)
+	if err != nil {
+		return 0
+	}
+	numDays := date.AddDate(0, 1, 0).Sub(date).Hours() / 24
+	return int(numDays)
+}
+
 func newApp() *App {
 	lastInvoiceDate := time.Now().Local().AddDate(0, 0, -1)
 	a := App{
@@ -435,6 +470,9 @@ func newApp() *App {
 	if a.IncludePreviousMonth {
 		a.invoiceMonths = append(a.invoiceMonths, a.lastInvoiceDate.AddDate(0, -1, 0).Format("2006-01"))
 	}
+	// The mandatory file saving period is the period since the first day of the previous month of last invoice date
+	previousMonthOfLastInvoiceDate := lastInvoiceDate.AddDate(0, -1, 0)
+	a.mandatoryFileSavingPeriodStartDate = time.Date(previousMonthOfLastInvoiceDate.Year(), previousMonthOfLastInvoiceDate.Month(), 1, 0, 0, 0, 0, previousMonthOfLastInvoiceDate.Location())
 
 	for _, month := range a.invoiceMonths {
 		a.filesToUpload[month] = make(map[string]struct{})
